# src/routing/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IErrorResponseDto": {
				"description": "Тело любого ответа с ошибкой (`status >= 400`).",
				"properties": {
					"status": {
						"type": "number",
						"format": "double",
						"description": "HTTP-статус."
					},
					"code": {
						"type": "string",
						"description": "Машинный код: доменный (`USER_EMAIL_TAKEN`) или по статусу (`NOT_FOUND`,\n`VALIDATION_ERROR`, `INTERNAL_ERROR`). Стабилен — на него опирается клиент."
					},
					"message": {
						"type": "string",
						"description": "Человекочитаемое сообщение; для 5xx — общее, без внутренних деталей."
					},
					"details": {
						"description": "Подробности: для VALIDATION_ERROR — поле → сообщение."
					},
					"requestId": {
						"type": "string",
						"description": "Идентификатор запроса — для обращения в поддержку и поиска в логах."
					},
					"stack": {
						"type": "string",
						"description": "Stack trace — только вне production."
					}
				},
				"required": [
					"status",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiKeyDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"prefix": {
						"type": "string",
						"description": "Открытая часть ключа: по ней ключ узнают в списке."
					},
					"scopes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"ownerId": {
						"type": "string"
					},
					"lastUsedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"revokedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"prefix",
					"scopes",
					"ownerId",
					"lastUsedAt",
					"expiresAt",
					"revokedAt",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreatedApiKeyDto": {
				"description": "Созданный ключ: `key` показывается один раз и больше не восстановим.",
				"properties": {
					"apiKey": {
						"$ref": "#/components/schemas/ApiKeyDto"
					},
					"key": {
						"type": "string",
						"description": "Полный ключ `<prefix>.<secret>` — передаётся в `X-Api-Key`."
					}
				},
				"required": [
					"apiKey",
					"key"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateApiKeyBody": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 100
					},
					"scopes": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Разрешения: `worker:demo.echo`, `worker:*`."
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"description": "Срок действия; без него ключ бессрочный."
					}
				},
				"required": [
					"name",
					"scopes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedDto_ApiKeyDto_": {
				"description": "Страница по смещению: списки с известным общим числом.",
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/ApiKeyDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double",
						"description": "Всего элементов, подходящих под фильтр."
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"total",
					"offset",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UUID": {
				"type": "string",
				"description": "Идентификатор сущности в пути или теле запроса. tsoa проверяет формат до\nконтроллера: неверный uuid → 422, а не ошибка Postgres.",
				"pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
			},
			"TPermission": {
				"type": "string",
				"description": "Право — строка `domain:action` (или `domain:*` — wildcard). Модуль объявляет\nсвои права сам через `definePermissions`; каталог с подписями отдаёт\n`GET /api/v1/permissions`."
			},
...
			"KnownRole": {
				"type": "string",
				"enum": [
					"admin",
					"user",
					"guest"
				],
				"description": "Тип для предопределённых ролей (автодополнение в IDE)."
			},
			"TRole": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/KnownRole"
					},
					{
						"allOf": [
							{
								"type": "string"
							},
							{
								"properties": {},
								"type": "object"
							}
						]
					}
				],
				"description": "Роль — произвольная строка; предопределённые значения дают автодополнение."
			},
			"IPermissionDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/TPermission"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRoleDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/TRole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserDto": {
				"description": "Пользователь для владельца и администрирования.",
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"emailVerified": {
						"type": "boolean"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"profile": {
						"$ref": "#/components/schemas/ProfileDto"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/IRoleDto"
						},
						"type": "array"
					},
					"directPermissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"email",
					"phone",
					"username",
					"roles",
					"directPermissions",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfileDto": {
				"description": "Профиль владельца.",
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"firstName": {
						"type": "string",
						"nullable": true
					},
					"lastName": {
						"type": "string",
						"nullable": true
					},
					"birthDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"gender": {
						"type": "string",
						"nullable": true
					},
					"locale": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/UserDto"
					}
				},
				"required": [
					"id",
					"userId",
					"firstName",
					"lastName",
					"birthDate",
					"gender",
					"locale",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IProfileUpdateRequestDto": {
				"properties": {
					"firstName": {
						"type": "string",
						"nullable": true,
						"maxLength": 40
					},
					"lastName": {
						"type": "string",
						"nullable": true,
						"maxLength": 40
					},
					"birthDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"gender": {
						"type": "string",
						"nullable": true,
						"maxLength": 20
					},
					"locale": {
						"type": "string",
						"nullable": true,
						"description": "Язык пользователя: `ru`, `en`, `en-US`. Письма — на `ru`/`en`, прочие\nязыки получают письма на `ru`.",
						"maxLength": 10
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PublicProfileDto": {
				"description": "Профиль глазами другого пользователя.",
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"firstName": {
						"type": "string",
						"nullable": true
					},
					"lastName": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"userId",
					"firstName",
					"lastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProfileListDto": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/PublicProfileDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double",
						"description": "Всего элементов, подходящих под фильтр."
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"total",
					"offset",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"apiKey": {
				"type": "apiKey",
				"in": "header",
				"name": "X-Api-Key",
				"description": "API-ключ: агент ноды (scope wg-agent) или интеграция"
			}
		}
	},
	"info": {
		"title": "wg-admin",
		"version": "1.0.1",
		"description": "WG Admin — админка WireGuard: бэкенд (Koa, tsoa, TypeORM, pg-boss) и агент нод на Go",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"paths": {
		"/api/v1/api-keys": {
			"post": {
				"operationId": "CreateApiKey",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreatedApiKeyDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Выпустить API-ключ сервиса. Полный ключ (`key`) возвращается только в\nэтом ответе — сохраните его: в БД хранится лишь хеш.",
				"summary": "Создание API-ключа",
				"tags": [
					"ApiKey"
				],
				"security": [
					{
						"jwt": [
							"permission:apikey:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateApiKeyBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListApiKeys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedDto_ApiKeyDto_"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Все API-ключи, новые первыми. Секреты не возвращаются.",
				"summary": "Список API-ключей",
				"tags": [
					"ApiKey"
				],
				"security": [
					{
						"jwt": [
							"permission:apikey:view"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/api-keys/{id}/revoke": {
			"post": {
				"operationId": "RevokeApiKey",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Отозвать ключ: запросы с ним сразу получают 401. Повторный отзыв — 204.",
				"summary": "Отзыв API-ключа",
				"tags": [
					"ApiKey"
				],
				"security": [
					{
						"jwt": [
							"permission:apikey:revoke"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
...
		"/api/v1/profile/my": {
			"get": {
				"operationId": "GetMyProfile",
				"responses": {
					"200": {
						"description": "Пользователь",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Получить профиль текущего пользователя.\nЭтот эндпоинт позволяет получить данные профиля пользователя, который выполнил запрос.\nИспользуется для получения информации о текущем пользователе, например, его имени, email, и других данных.",
				"summary": "Получение профиля текущего пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/profile/my/update": {
			"patch": {
				"operationId": "UpdateMyProfile",
				"responses": {
					"200": {
						"description": "Обновленный профиль пользователя",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Обновить профиль текущего пользователя.\nЭтот эндпоинт позволяет пользователю обновить свои данные, такие как имя, email и другие параметры профиля.",
				"summary": "Обновление профиля текущего пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Обновленные данные профиля",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileUpdateRequestDto",
								"description": "Обновленные данные профиля"
							}
						}
					}
				}
			}
		},
		"/api/v1/profile/my/delete": {
			"delete": {
				"operationId": "DeleteMyProfile",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Очистить профиль текущего пользователя.\nЛичные данные (имя, фамилия, дата рождения, пол) обнуляются,\nсама запись профиля остаётся.",
				"summary": "Очистка профиля текущего пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/profile/all": {
			"get": {
				"operationId": "GetProfiles",
				"responses": {
					"200": {
						"description": "Страница профилей",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileListDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Получить все профили постранично, новые первыми.",
				"summary": "Получение всех профилей",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": [
							"permission:profile:view"
						]
					}
				],
				"parameters": [
					{
						"description": "Смещение, по умолчанию 0",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Размер страницы: по умолчанию 20, не больше 100",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/profile/{userId}": {
			"get": {
				"operationId": "GetProfileById",
				"responses": {
					"200": {
						"description": "Пользователь по ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PublicProfileDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Получить профиль по ID.\nЭтот эндпоинт позволяет получить профиль другого пользователя по его ID. Доступен только для администраторов.",
				"summary": "Получение профиля по ID",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, профиль которого нужно получить",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/api/v1/profile/update/{userId}": {
			"patch": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Обновленный профиль пользователя",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Обновить профиль другого пользователя.\nЭтот эндпоинт позволяет администраторам обновлять профиль других пользователей.",
				"summary": "Обновление профиля другого пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": [
							"permission:profile:update"
						]
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, профиль которого необходимо обновить",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"description": "Данные для обновления профиля",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileUpdateRequestDto",
								"description": "Данные для обновления профиля"
							}
						}
					}
				}
			}
		},
		"/api/v1/profile/delete/{userId}": {
			"delete": {
				"operationId": "DeleteProfile",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Очистить профиль другого пользователя.\nЛичные данные обнуляются, запись профиля остаётся.",
				"summary": "Очистка профиля другого пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": [
							"permission:profile:delete"
						]
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, профиль которого необходимо очистить",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		}
	]
}